Reduce client usage
